Properly handle free & realloc for initial memory
Some memory allocated by runtime libraries and used throughout the
execution may still be coming from the initial malloc pool. When such a
pointer is passed to free, there is nothing to do. However, if it were
to be passed to realloc, we need to allocate a block of memory using
malloc and copy the contents of the original buffer to that brand new
memory region.